deployment/requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
pillow
matplotlib
pandas
